+++ demo/node/index.js
@@ -13,6 +13,9 @@ const test = async () => {
         familyName: 'Doe',
         birthYear: 1971,
         usCitizen: true,
+        avatar: Buffer.from(
+          '<?xml version="1.0" standalone="no"?><svg viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg"><circle cx="50" cy="50" r="50" /></svg>'
+        ),
       },
       name: 'my personal data',
     })