+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] (2023-09-20)
+
+### Changed
+
+- Improved error handling
+- The `protectedData` param in `fetchGrantedAccess` is no longer required and now accepts `"any"`, when not specified the method will fetch granted access for any protectedData
+- Removed bundle
+- Updated iexec to version 8.2.1
+- Updated supported MIME types (not supported will fallback to `application/octet-stream`):  
+  - application/octet-stream
+  - application/pdf
+  - application/xml
+  - application/zip
+  - audio/midi
+  - audio/mpeg
+  - audio/x-wav
+  - image/bmp
+  - image/gif
+  - image/jpeg
+  - image/png
+  - image/webp
+  - video/mp4
+  - video/mpeg
+  - video/x-msvideo
+
 ## [0.1.1]
 
 ### Changed